Finished Emotional Check In
Emotional Checkin Works as designed
