+++ package.json
@@ -10,13 +10,17 @@
   },
   "dependencies": {
     "@expo-google-fonts/dm-sans": "^0.3.0",
+    "@react-navigation/native": "^7.1.9",
+    "@react-navigation/native-stack": "^7.3.13",
     "expo": "~53.0.9",
     "expo-font": "~13.3.1",
     "expo-linear-gradient": "~14.1.4",
     "expo-status-bar": "~2.2.3",
     "react": "19.0.0",
     "react-native": "0.79.2",
-    "react-native-svg": "15.11.2"
+    "react-native-svg": "15.11.2",
+    "react-native-screens": "~4.10.0",
+    "react-native-safe-area-context": "5.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
